requirements/requirements-dev.txt
Django==3.1
djangorestframework==3.11.0
django-environ==0.4.5
django-filter==2.3.0
pytest==6.0.1
pytest-django==3.9.0
psycopg2